Split out utility function from digest.ts, allow for different email routes
