+++ server/routes/email.tsx
@@ -4,41 +4,15 @@ import ReactDOMServer from 'react-dom/server';
 import { ServerStyleSheet, StyleSheetManager } from 'styled-components';
 import juice from 'juice';
 import { handleErrors } from 'server/utils/errors';
-import pick from 'lodash.pick';
-import util from 'util';
-import omit from 'lodash.omit';
-import { GroupedActivityItems } from 'client/components/Email/Digest';
 
 import { hostIsValid } from 'server/utils/routes';
 import { getInitialData } from 'server/utils/initData';
 import { minify } from 'html-minifier';
 
-import { Digest } from 'components/Email';
 import { reset, globals } from 'components/Email/styles';
-import { ActivityItem } from 'types';
-import { fetchActivityItems } from 'server/activityItem/fetch';
+import { Digest } from 'components/Email';
 
-const getAffectedObject = (item) =>
-	item.pubId
-		? { id: item.pubId, title: item.payload.pub.title }
-		: item.collectionId
-		? { id: item.collectionId, title: item.payload.collection.title }
-		: 'page' in item.payload
-		? pick(item.payload.page, ['id', 'title'])
-		: { id: item.communityId, title: item.payload.community.title };
-
-const getAffectedObjectIcon = (item: ActivityItem) =>
-	item.pubId
-		? 'pubDoc'
-		: item.collectionId
-		? 'collection'
-		: 'page' in item.payload && item.payload.page.id
-		? 'page-layout'
-		: 'office';
-
-type KeyedActivityItem = ActivityItem & {
-	displayKey: string;
-};
+import { getDigestData } from 'server/utils/email';
 
 const inlineStylesWithMarkup = (emailMarkup: React.ReactNode, extraStyles: string) => {
 	const stylesheet = new ServerStyleSheet();
@@ -75,62 +49,35 @@ export const render = (emailMarkup: React.ReactNode, extraStyles = '') => {
 	)}</html>`;
 };
 
-app.get('/email', async (req, res, next) => {
+const templates = {
+	digest: {
+		prepData: getDigestData,
+		component: Digest,
+	},
+};
+
+app.get('/email/:templateSlug', async (req, res, next) => {
 	try {
 		const initialData = await getInitialData(req, true);
 		const {
-			communityData: { id: communityId },
-			scopeData: { scope },
+			communityData: community,
 			loginData: { id: userId },
 		} = initialData;
-		if (!hostIsValid(req, 'community') || process.env.NODE_ENV === 'production' || !userId) {
+		const { templateSlug } = req.params;
+		if (
+			!hostIsValid(req, 'community') ||
+			process.env.NODE_ENV === 'production' ||
+			!userId ||
+			!(templateSlug in templates)
+		) {
 			return next();
 		}
-		const { activityItems, associations } = await fetchActivityItems({ scope });
-		const activityItemsGroupedByObjectId: Record<
-			string,
-			KeyedActivityItem[]
-		> = activityItems.reduce((result, item) => {
-			const objectId = getAffectedObject(item).id;
-			const payloadKeys = Object.keys(item.payload)
-				.sort()
-				.join();
-			const displayKey = `${item.kind} - ${payloadKeys}`;
-			return {
-				...result,
-				[objectId]: [...(result[objectId] || []), { ...item, displayKey }],
-			};
-		}, {});
-		const dedupedActivityItems = Object.keys(activityItemsGroupedByObjectId).reduce(
-			(memo, objectId) => ({
-				...memo,
-				[objectId]: {
-					items: activityItemsGroupedByObjectId[objectId]
-						.sort((first, second) => (first.timestamp > second.timestamp ? -1 : 1))
-						.reduce(
-							(result, item) =>
-								item.displayKey in result
-									? result
-									: { ...result, [item.displayKey]: item },
-							{},
-						),
-					title: getAffectedObject(activityItemsGroupedByObjectId[objectId][0])?.title,
-					icon: getAffectedObjectIcon(activityItemsGroupedByObjectId[objectId][0]),
-				},
-			}),
-			{},
-		);
-
-		const communityItems: GroupedActivityItems = pick(dedupedActivityItems, communityId);
-		const pubItems: GroupedActivityItems = omit(dedupedActivityItems, communityId);
+		const { component, prepData } = templates[templateSlug];
 		return res.send(
 			render(
-				Digest({
-					userId, // replace this with email recipient's userId
-					community: initialData.communityData,
-					associations,
-					pubItems,
-					communityItems,
+				component({
+					community,
+					...prepData(initialData),
 				}),
 			),
 		);
