+++ server/utils/email/reset.ts
@@ -23,21 +23,3 @@ export const sendPasswordResetEmail = ({ toEmail, resetUrl }) => {
 		`),
 	});
 };
-
-export const sendSignupEmail = ({ toEmail, signupUrl }) => {
-	return mg.messages.create('mg.pubpub.org', {
-		from: 'PubPub Team <hello@pubpub.org>',
-		to: [toEmail],
-		subject: 'Welcome to PubPub!',
-		text: stripIndent(`
-			Welcome to PubPub!
-
-			Click the following link to create your account:
-
-			${signupUrl}
-
-			Sincerely,
-			PubPub Support
-		`),
-	});
-};
